refactor: removed old files and references, updated docs testing
