+++ .github/workflows/test.yml
@@ -30,4 +30,4 @@ jobs:
 
             - name: Check if repo apidoc is up to date
               run: |
-                git diff-index --quiet HEAD
+                git diff --exit-code HEAD
