--- docs/api/tetris.impl.modern.rst
@@ -1,6 +1,0 @@
-tetris.impl.modern module
-=========================
-
-.. automodule:: tetris.impl.modern
-   :members:
-   :show-inheritance:
