+++ docs/api/tetris.impl.presets.rst
@@ -1,0 +1,6 @@
+tetris.impl.presets module
+==========================
+
+.. automodule:: tetris.impl.presets
+   :members:
+   :show-inheritance:
