+++ tox.ini
@@ -3,7 +3,7 @@ isolated_build = true
 envlist = py39, py310
 
 [testenv]
-whitelist_externals = poetry
+allowlist_externals = poetry
 deps =
     flake8
     isort
@@ -19,10 +19,12 @@ commands =
 
 [testenv:docs]
 description = generate apidocs using sphinx
+allowlist_externals = rm
 deps = sphinx
 setenv =
     SPHINX_APIDOC_OPTIONS = members,show-inheritance
 commands =
+    rm -rf docs/api
     sphinx-apidoc -Tefo docs/api tetris
 
 [flake8]
